Use UUID strings for versions instead of incremental numbers (#8)
This changes Versions to contain opaque strings instead of incremental numbers. We define a format for version strings that contains a timestamp and a UUID. We also remove the nextVersion method on the TableVersion interface as clients can now generate their own Versions as needed.

This means that written partitions will be completely immutable as code will generate new Versions every time it attempts to write data. It also ensures two concurrent clients can't attempt to write to the same path.


Co-authored-by: Patricio Vighi <pvighi@gmail.com>