Metastore refactorings (#25)
This includes the following refactoring changes:

- Remove metastore module, moving Metastore base trait and related code into core.
This base trait is as 'core' as TableVersions so it doesn't make sense to have a separate module for it. Concrete implementations will still live in other modules, of course.

- Add VersionedMetastore trait composed of TableVersions and Metastore traits.
Moving common functionality that can be implemented purely in terms of the underlying traits into this trait.